FF7: Refactor field.cpp into multiple files
Created a new namespace for all the files related to field.

The new list of modules/files is:
 - background: for field background related stuff
 - enter: related to `field_enter` sub run only once when transitioning to FIELD mode
 - model: for field models related stuff
 - opcode: related to field opcodes
 - utils: common or generic utils related to multiple modules
 - defs: functions that are exported outside of ff7_field namespace
 - field: main file that contains unrelated stuff to the other files and the hook init function
